remove closing slash in img tag
As @zegnat pointed out in dissolve/mf2-tester#4, the html value should
follow the HTML spec for serializing HTML fragments, which does not
include the trailing slash on self-closing tags.

There is an open question as to how strictly the test suite should
enforce this, given that this is often handled by underlying HTML
libraries outside the microformats parser's control.  But regardless,
the test suite itself should follow the spec as strictly as possible.