Reflect the parsing rules for src and alt attributes on img elements
Parsing a `u-` property should return the result of "parse an img
element for src and alt" in case the element is a `img` tag with a
`src` attribute. Based on this sub section `img` tags with a `alt`
attribute should result in a new structure, containing a `value` and
`alt` property.
> http://microformats.org/wiki/microformats2-parsing

This mr updates all tests subject to this rule in order to reflect this.

!!! This mr does not include any updates to the v1 tests nor did i
check if any changes to are needed, as i'm not be familiar with v1
parsing rules !!!